corrigir erros

+++ Grupo2A/Model/DTO/TipoDePrato2listing_dto.cs
@@ -1,7 +1,7 @@
 namespace Cozinha_BE.Model.DTO;
 
 public class TipoDePrato2listing_dto{
-    public int IdTipoPrato {get; set; }
+    public long IdTipoPrato {get; set; }
     public required string Nome {get; set;}
     public required string DescricaoTipoPrato {get; set;}
 }
