+++ Grupo2A/Model/TipoDePrato.cs
@@ -5,7 +5,7 @@ namespace Cozinha_BE.Model
     public class TipoDePrato
     {
         [Key]
-        public int IdTipoPrato { get; set; } // Identificador único do tipo de prato
+        public long IdTipoPrato { get; set; } // Identificador único do tipo de prato
 
         [StringLength(50)]
         public required string Nome { get; set; } // Nome do tipo de prato (ex: Carne, Peixe, Vegetariano)
